fix bugs 1 qc module fix coverage over 100% because vector resize method resize the container size not change elements in container
fix features
input parameter change sample_id(i) to read_group(g)

+++ qc_report/qc_result_analysis.h
@@ -23,7 +23,7 @@ namespace gamtools {
             std::string chr_name;
             int tid;
             int lens;
-            int bases;
+            int64_t bases;
             int cover_pos;
             double cover_precent;
             double mean_depth;
